v2_automated_thematic_search/requirements.txt
aiohttp>=3.9
biopython>=1.84
feedparser>=6.0
numpy>=1.26
openpyxl>=3.1
pandas>=2.2
PyYAML>=6.0
requests>=2.31
sentence-transformers>=3.0
tqdm>=4.66
